correção nos tipos de dados numericos

+++ src/main/java/dio/bootcamp/java/api_rest_spring_boot/model/Card.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Entity(name = "tb_card")
@@ -14,5 +16,5 @@ public class Card {
     @Column(unique = true)
     private String number;
     @Column(name = "available_limit", scale = 12, precision = 12)
-    private Double limit;
+    private BigDecimal limit;
 }
